Split and merge NALUs in their original format
The decoder requires the frame data to follow the extradata format (avcC/hvcC -> length-prefixed, Annex B -> Annex B). To avoid having to rewrite the extradata, we split and merge NALUs in their original format.

Once again, I expect to remove all of this when we can use bitstream filters in libav.js